Fix regression in manager selection (v1.1.1)
The default behavior was to select a _default_manager, which causes
issues as Armstrong's base content model does not use `objects` as its
default manager.  This reverts the code back to its original behavior
from 1.x.  We will revisit once `armstrong.core.arm_content` is at 2.0.0
and can expose the correct manager as the default.